RCA 리팩토링
