+++ cat-jjal-maker-rca/src/components/title.js
@@ -1,0 +1,2 @@
+const Title = ({ children }) => (<h1>{children}</h1>);
+export default Title;
